+++ dakara_player_vlc/text_generator.py
@@ -3,13 +3,9 @@ import logging
 from codecs import open
 from configparser import ConfigParser
 
-from jinja2 import Environment, FileSystemLoader
+from jinja2 import Environment, FileSystemLoader, ChoiceLoader
 
-
-SHARE_DIR = 'share'
-SHARE_DIR_ABSOLUTE = os.path.join(os.path.dirname(os.path.abspath(__file__)),
-                                  os.pardir,
-                                  SHARE_DIR)
+from dakara_player_vlc.resources_manager import get_file, PATH_TEMPLATES
 
 
 TRANSITION_TEMPLATE_NAME = "transition.ass"
@@ -57,8 +53,11 @@ class TextGenerator:
     def load_templates(self, config):
         # create Jinja2 environment
         self.environment = Environment(
-                loader=FileSystemLoader(SHARE_DIR_ABSOLUTE)
-                )
+            loader=ChoiceLoader([
+                FileSystemLoader(config.get('templateDirectory', '')),
+                FileSystemLoader(PATH_TEMPLATES)
+            ])
+        )
 
         # add filter for converting font icon name to character
         self.environment.filters['icon'] = self.convert_icon
@@ -68,16 +67,12 @@ class TextGenerator:
             self.convert_link_type_name
         )
 
-        transition_template_path = config.get(
-                'transitionTemplateName',
-                TRANSITION_TEMPLATE_NAME
-                )
-
-        idle_template_path = config.get('idleTemplateName', IDLE_TEMPLATE_NAME)
-
         # load templates
-        self.load_transition_template(transition_template_path)
-        self.load_idle_template(idle_template_path)
+        self.load_transition_template(config.get('transitionTemplateName',
+                                                 TRANSITION_TEMPLATE_NAME))
+
+        self.load_idle_template(config.get('idleTemplateName',
+                                           IDLE_TEMPLATE_NAME))
 
     def convert_icon(self, name):
         """Convert the name of an icon to its code
@@ -148,69 +143,68 @@ class TextGenerator:
     def load_icon_map(self):
         """ Load the icon map
         """
-        icon_map_path = os.path.join(SHARE_DIR_ABSOLUTE, ICON_MAP_FILE)
-
-        if not os.path.isfile(icon_map_path):
-            raise IOError("Icon font map file '{}' not found".format(
-                icon_map_path
-                ))
+        icon_map_path = get_file(ICON_MAP_FILE)
 
         icon_map = ConfigParser()
         icon_map.read(icon_map_path)
         self.icon_map = icon_map['map']
 
-    def load_transition_template(self, template_name):
+    def load_transition_template(self, transition_template_name):
         """ Load transition screen text template file
 
             Load the default or customized ASS template for
             transition screen.
         """
-        template_path = os.path.join(SHARE_DIR_ABSOLUTE, template_name)
-        template_default_path = os.path.join(SHARE_DIR,
-                                             TRANSITION_TEMPLATE_NAME)
+        loader_custom, loader_default = self.environment.loader.loaders
 
-        if os.path.isfile(template_path):
-            pass
+        if transition_template_name in loader_custom.list_templates():
+            logger.debug(
+                "Loading custom transition template file '{}'".format(
+                    transition_template_name
+                )
+            )
 
-        elif os.path.isfile(template_default_path):
-            logger.warning("Transition template file not found \"{}\", \
-using default one".format(template_path))
+            self.transition_template = self.environment.get_template(
+                transition_template_name)
 
-            template_name = TRANSITION_TEMPLATE_NAME
+            return
 
-        else:
-            raise IOError("No template file for transition screen found")
+        if TRANSITION_TEMPLATE_NAME in loader_default.list_templates():
+            logger.debug("Loading default transition template file")
 
-        logger.debug("Loading transition template file \"{}\"".format(
-            template_path
-            ))
+            self.transition_template = self.environment.get_template(
+                TRANSITION_TEMPLATE_NAME)
 
-        self.transition_template = self.environment.get_template(template_name)
+            return
 
-    def load_idle_template(self, template_name):
+        raise IOError("No template file for transition screen found")
+
+    def load_idle_template(self, idle_template_name):
         """ Load idle screen text template file
 
             Load the default or customized ASS template for
             idle screen.
         """
-        template_path = os.path.join(SHARE_DIR_ABSOLUTE, template_name)
-        template_default_path = os.path.join(SHARE_DIR_ABSOLUTE,
-                                             IDLE_TEMPLATE_NAME)
+        loader_custom, loader_default = self.environment.loader.loaders
 
-        if os.path.isfile(template_path):
-            pass
+        if idle_template_name in loader_custom.list_templates():
+            logger.debug(
+                "Loading custom idle template file '{}'".format(
+                    idle_template_name
+                )
+            )
 
-        elif os.path.isfile(template_default_path):
-            logger.warning("Idle template file not found \"{}\", \
-using default one".format(template_path))
+            self.idle_template = self.environment.get_template(
+                idle_template_name)
 
-            template_name = IDLE_TEMPLATE_NAME
+            return
 
-        else:
-            raise IOError("No template file for idle screen found")
+        if IDLE_TEMPLATE_NAME in loader_default.list_templates():
+            logger.debug("Loading default idle template file")
 
-        logger.debug("Loading idle template file \"{}\"".format(
-            template_path
-            ))
+            self.idle_template = self.environment.get_template(
+                IDLE_TEMPLATE_NAME)
 
-        self.idle_template = self.environment.get_template(template_name)
+            return
+
+        raise IOError("No template file for idle screen found")